typo python

+++ devel/python3/01-python-3.9.1.sh
@@ -2,7 +2,7 @@
 ${log} `basename "$0"` " started" blfs_all &&
 
 ${log} `basename "$0"` " download" blfs_all &&
-# if test -d /sources/Python-3.9.1
+if test -d /sources/Python-3.9.1
  then
   rm -rf /sources/Python-3.9.1
 fi
